tests: move tests for create signed read URLs to new module
Toward #475
