+++ tests/system/_helpers.py
@@ -25,6 +25,7 @@ retry_429_503 = RetryErrors(
 )
 
 user_project = os.environ.get("GOOGLE_CLOUD_TESTS_USER_PROJECT")
+signing_blob_content = b"This time for sure, Rocky!"
 
 
 def _bad_copy(bad_request):
